MAP/requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-filter==21.1
djangorestframework==3.12.4
idna==2.10
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.3
xmltodict==0.12.0
